backend/requirements.txt
appdirs==1.4.4
asgiref==3.2.10
astroid==2.4.2
autopep8==1.5.4
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
distlib==0.3.1
Django==3.1
django-filter==2.4.0
django-memcached==0.1.2
django-multiselectfield==0.1.12
django-phone-field==1.8.1
djangorestframework==3.12.2
filelock==3.0.12
futures==3.1.1
goslate==1.5.1
idna==2.10
isort==5.4.2
lazy-object-proxy==1.4.3
Markdown==3.3.3
mccabe==0.6.1
nose==1.3.7
pbr==5.5.1
pycodestyle==2.6.0
PyDictionary==2.0.1
pylint==2.6.0
pytz==2020.1
Random-Word==1.0.4
requests==2.24.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
stevedore==3.2.2
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.10
virtualenv==20.1.0
virtualenv-clone==0.5.4
virtualenvwrapper==4.8.4
wrapt==1.12.1
